codex/requirements.txt
streamlit>=1.36.0
pandas>=2.2.2
plotly>=5.22.0
kiteconnect>=5.0.1
python-dotenv>=1.0.1
apscheduler>=3.10.4
pydantic>=2.8.2
PyYAML>=6.0.2
numpy>=2.0.0
